+++ src/Support/ValidatesWebData.php
@@ -1,9 +1,10 @@
 <?php
 
-namespace SergiX44\Nutgram\Web;
+namespace SergiX44\Nutgram\Support;
 
-use SergiX44\Nutgram\Web\Entities\LoginData;
-use SergiX44\Nutgram\Web\Entities\WebAppData;
+use SergiX44\Nutgram\Exception\InvalidDataException;
+use SergiX44\Nutgram\Telegram\Web\LoginData;
+use SergiX44\Nutgram\Telegram\Web\WebAppData;
 
 trait ValidatesWebData
 {
@@ -83,39 +84,5 @@ trait ValidatesWebData
 
         // return the hash and the sorted data
         return [$hash, $stringData];
-    }
-
-    /**
-     * Generates webapp data + hash.
-     * @param array $data The data to generate webapp data from.
-     * @return string The generated webapp data as query string.
-     * @internal For testing purposes only.
-     */
-    public function generateWebAppData(array $data): string
-    {
-        $queryString = http_build_query(array_filter($data));
-
-        [, $sortedData] = $this->parseQueryString($queryString);
-        $secretKey = $this->createHashHmac($this->token, 'WebAppData');
-        $hash = bin2hex($this->createHashHmac($sortedData, $secretKey));
-
-        return $queryString.'&hash='.$hash;
-    }
-
-    /**
-     * Generates login data + hash.
-     * @param array $data The data to generate login data from.
-     * @return string The generated login data as query string.
-     * @internal For testing purposes only.
-     */
-    public function generateLoginData(array $data): string
-    {
-        $queryString = http_build_query(array_filter($data));
-
-        [, $sortedData] = $this->parseQueryString($queryString);
-        $secretKey = $this->createHash($this->token);
-        $hash = bin2hex($this->createHashHmac($sortedData, $secretKey));
-
-        return $queryString.'&hash='.$hash;
     }
 }
